Updated the 1.0 to 1.1 database script to correctly split pends
I wrapped the whole script in a transaction so that a failure in the
middle will not leave the DB in an inconsistant state. As I result I
needed to remove the VACUUM call and disable and reenable triggers.

I mistakenly omitted table ownership changes in the previous version of
the script.

I had mistakenly set the treemap_plotpending.geometry column as required.
A plotpending can contain a non-geo field. This required I make a
complementary change to the PlotPending model.

I made lots of syntax errors in the fill_treemap_pending function that
are corrected.