Start converting the application to use the new Plot model.
Moving the geometry field to the Plot has far reaching implications.
In this commit I have made the updates required to get the main map page
to load without error, to have the tree detail and edit pages load without
errors and to show a detail popup without errors when a tree is clicked.

Many features are still broken.