Moved Geocoder contstruction settings out of settings.py
Contructing the Geocoder in the implementation settings file was crashing
when deployed because the WSGI file needs to load the virtual environment
path from the settings file before the omgeo module is available.

I now take advantage of the fact that an omgeo Geocoder can be initialized
with an array of sources specified by strings, avoiding the chicken and egg
problem.