Sync key names in the plot JSON representation with Plot form
This should have been included with 416a9d45

Fields were not being loaded and saved correctly in the iOS app because
there were small inconsistancies in the key naming.

I also removed the depreciated 'field_whitelist' variable and moved the tree
field whitelist variable declaration closer to where it is used in the
function.