Update api to create pends
When the pending system is enabled, a public user who edits a plot or a tree
creates pending edit records to be approved by a manager.

To implement this I added created_by and was_created_by_a_manager properties
to the Tree and Plot models (via a mixin) and then used those properties in
the api views to determine whether or not to create pending records.

I also added a helpful set_create_attributes method to the Pending model
and the TreePending and PlotPending subclasses to encapsulate setting the
proper attribute values on a new pend, which keeps that logic out of the view.