Fix to update_plot view to correctly handle pending
The view determines whether to create a pend or a real edit by checking the
audit trail of the plot for a record of type 'I' (insert). An upgraded
database may not have audit insert records for plots, so I am assuming to
treat the plot as if it was create by and administrator if there is no
insert audit record.