Add API endpoints for approving and rejecting pending edits
I discussed this functionality with Carissa and we decided that the original
pending view and model should be updated so that approving an edit for
a field rejects all other pending edits for that field.

I added this new 'approve and reject' functionality into TreePending and
PlotPending models and make use of it in both the treemap and api views.

To simplify maintenance I tried to keep the api and treemap views as small
as possible and to share as much code between them as possible.