Add plot and tree delete methods to the API
The mobile app needs endpoints for removing plots and trees.

To avoid duplication I moved the existing delete code used by the treemap
application into the Plot and Tree models and call it from both the
treemap application and the api.

I needed a /plots/{plot_id}/tree endpoint to send a DELETE command to so,
for completeness, I added a GET view function which simply returns the value
of the 'tree' key from the /plots/{plot_id} response.