Update new plot API to return full tree JSON
An interesting bug popped up in this situation

- Filter the trees
- Select a filtered tree
- Edit the tree so that it no longer matches the filter

Previously the map view conroller responded to a successful edit by
calling the same method that handeld a user tap, "select nearest plot"
Because this plot no longer matched the filter, the selection jumpped
to a different plot, which was jarring.

To keep the mobile code simple I am now returning the full plot/tree
detail in the 201 Created response so that the UI can be updated without
making a second HTTP request.