Fix issue with missing species filter and clicking
Clicking with the missing species filter applied selected some
plots that had a species set. The root cause was that the
plots with trees that had been marked 'present=False' (deleted)
were co-opting the query.

The fix was to force the query to make sure that a null
plot was found only on the present tree.