Rename Tree.delete()/Plot.delete() to *.remove()
I unthinkingly stomped on the Django Model.delete() function when I pushed
delete functionality from the views into the models. Although the application
never truly deletes trees and plots, it is important that the models behave
correctly if a unit test or management task calls .delete() expecting the
row to be removed from the database.