Added catch and passthrough for distance searches
The distance search function takes care of the geographical element of searching. The search function used to be
arranged such that the distance searches didn't throw geographical searches. In the course of refactoring, we
closed that hole without realizing it, so distance searches were only happening within the neighborhood of the
starting point. Added a pass to the search function so distance searches are now working again.