Revert changes from 4c58f6d1119ae
The newly added `recently_edited_plots` method uses
fields that do not actually exist in the database.

Since I can't be 100% sure what these fields are actually for,
and the fact there is only one date field on a plot, I've
replaced the offending field (date_planted) with `last_updated`.