Refactor searching
Previously the secret _build_tree_search_result was a 600+
line method of very dense and sort of crazy python

The new approach offloads most of the searching code into
search.py

search.py is essentially a set of "apply_X_filter(req,searchObj)"
methods.

When apply_filters(...) is called, it iterates through each filter
and applies it in order.

There are still future things that need to be done:
* It sucks that we have to repeat ourselves with the plot__/tree__
  stuff everywhere

* Each "apply_X_filter" method should probably have its own test
  (note that most are already covered by the search integration
   tests)

* Stewardship stuff still scares me a bit

* The "append" methods called on tile_query should instead
  create new arrays. This would gives us true referential
  transparency