Fix failing test by force reloading form field choices
The choice field options in treemap.TreeAddForm are defined when the
form module is loaded and never again. When the unit test changes the CHOICES
setting, the change was not being reflected in the form.

I attempted to reload the entire treemap.forms module, but this
caused an exception to be thrown when the __init__ method of TreeAddForm
calls the super class __init__. My solution, instead, was to
force update the options in the form field.