Update api to correctly handle non-decimal version headers
The choice conversion logic raising exceptions when the client passed
an ApplicationVersion header with a non-numeric version number, including
'1.2.3' which is a valid version number. I replaced Decimal parsing with
StrictVersion from distutils and added a check that assumes that if
the version string is not parsable, that the app making the API request
is an old version.