Fix tree_count assignment on species save
On save, tree_count was assigned to the count of trees in the
tree_set property of the species model. This doesn't work for
a species that has no pk, as in the case of new species.

Stop this problem by only updating this count when pk == True

tree_count defaults to zero for new species, so the value will
be correct without this update anyway.