Support for additional species fields
In particular, a "family" field and a "other_part_of_name" field.

The other_part field contains any sort of text that the users want and
are used by the uploader to match species accordingly. For instance, if
someone wants to use a varietal or a sub-species this is the correct
place for it.