Fix Ban/Activate links that contain duplicate DOM ids
This bug was reported as occurring when the the ban/activate links are
clicked after sorting or filtering:

""" On the users page (http://treezilla.org/users/), administrators
can use the ban? button to the far right of a user name to ban the
person from logging into OTM. If you click the ban button, it becomes
an activate button. If you click the activate button, it makes the
person's inactive account (meaning they never clicked the activation
link in the email) become active.

The problem is that the ban/activate button isn't working if you do
any type of sorting or filtering. It will work on the first page of
user results but any action to change the search results (clicking on
a column header), changing pages, etc makes the button no longer work.
"""

However, it looks like the problem was actually that for some, but not
all users, the DOM element that their data and links lived inside of
contained IDs that also occurred in the species list. jQuery selection
was not working because this was causing invalid HTML. This problem is
fixed by changing the user DOM element to have a class and a data-id
instead of an id attribute.