Fix Ban/Activate displayed option for user edits
The problem was that for users who were labeled "Inactive", a "Ban?"
link was displaying, even though it should have said "Activate?" -
banning is only a valid option for users that are labeled
"Active". The admin was required to first click "Ban?", which would
save the user object without changing it, then click "Activate?" to
produce the desired result.

This is fixed by changing how the initial html template is rendered.