Disable zoomMethod and transitionEffect
This regression was introduced by c589fea69d8e57bd87e61fe762223a4251f8bff1, upgrading to version 2.13 of OpenLayers, in order to fix broken zoom on mobile devices.

Dots from pervious zoom levels were lingering after zooming, which was determined to be caused by the way OpenLayers was replacing old tiles with new ones in conjunction with slowdowns in geoserver and/or lag in getting tiles from the google maps api.

This fix disables the zoomMethod and transitionEffect, both of with work together to control the animation of how layers change. They perform much faster now.

The only wrinkle here, is that transitionEffect is not documented for the version of OpenLayers we're using (2.13), it's from the 2.6 docs. It's default for 2.6 is listed as null, the same value assignment that fixed this bug. This points to the fact that there's some underdocumented new feature addition or old feature deprecation. It's something to watch out for if this breaks again.