Force CONTACT_EMAILS into a tuple
to fix a mutable state bug. settings is an immutable object, but settings.CONTACT_EMAILS was a list. Over the lifetime of the appserver process, this list was being mutated to contain ALL senders. This was causing all users that have used the contact form and requested a copy, to get EVERYONE's copy.

Conflicts:

	treemap/views.py