Add management cmd for exporting audits to OTM2
For standard models like Plot and Tree, records are exported using the standard django 'dumpdata' command, and imported using a custom fixture loader with a small DSL for field/value transformation rules.

OTM1 audits differ from OTM2 audits to an extent much greater than that of plots or trees. For this reason, we'll use a custom exporter to produce a fixture-like data dump that the OTM2 custom loader is able to understand.