Add performant dumpdata alternative
A quick and dirty solution to handle the problem of django running out of memory when dumping JSON fixtures.

This was fixed in Django 1.6 (see below), but this commit is a simpler way to provide a new command that plucks just the features we need, instead of a proper backport.
https://github.com/django/django/commit/3b5083bee5e96539dec599106aece9889e70ce05#diff-7a82040e56038c35d815475a068d481dR57